restructure the folders 2
