+++ src/main/java/logic/commands/MemberDeleteCommand.java
@@ -1,11 +1,11 @@
-package commands;
+package logic.commands;
 
-import members.Member;
-import tasks.Task;
-import utils.CommandResult;
+import model.members.Member;
+import model.tasks.Task;
+import logic.CommandResult;
 import utils.DukeException;
 import utils.Storage;
-import core.Ui;
+
 import java.util.ArrayList;
 
 public class MemberDeleteCommand extends Command {
