+++ src/main/java/logic/commands/RecurringCommand.java
@@ -1,10 +1,9 @@
-package commands;
+package logic.commands;
 
-import members.Member;
-import core.Ui;
-import tasks.Event;
-import tasks.Task;
-import utils.CommandResult;
+import model.members.Member;
+import model.tasks.Event;
+import model.tasks.Task;
+import logic.CommandResult;
 import utils.DukeException;
 import utils.Storage;
