+++ src/main/java/logic/parsers/DoneCommandParser.java
@@ -1,7 +1,7 @@
-package parsers;
+package logic.parsers;
 
-import commands.Command;
-import commands.DoneCommand;
+import logic.commands.Command;
+import logic.commands.DoneCommand;
 import utils.DukeException;
 
 public class DoneCommandParser {
