+++ src/main/java/logic/parsers/SnoozeCommandParser.java
@@ -1,7 +1,7 @@
-package parsers;
+package logic.parsers;
 
-import commands.Command;
-import commands.SnoozeCommand;
+import logic.commands.Command;
+import logic.commands.SnoozeCommand;
 
 public class SnoozeCommandParser {
     public static Command parse(String userInput) {
