+++ src/main/java/logic/parsers/UnlinkCommandParser.java
@@ -1,7 +1,7 @@
-package parsers;
+package logic.parsers;
 
-import commands.Command;
-import commands.UnlinkCommand;
+import logic.commands.Command;
+import logic.commands.UnlinkCommand;
 
 public class UnlinkCommandParser {
     public static Command parse(String userInput) {
