+++ src/main/java/model/tasks/Task.java
@@ -1,4 +1,4 @@
-package tasks;
+package model.tasks;
 
 import utils.DukeException;
