+++ src/test/java/parser/ArgumentTokenizerTest.java
@@ -1,14 +1,9 @@
 package parser;
 
 import org.junit.jupiter.api.Test;
-import parsers.ArgumentTokenizer;
-import parsers.DukeParser;
-import parsers.SpellingErrorCorrector;
+import logic.parsers.ArgumentTokenizer;
 import utils.DukeException;
 
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.Date;
 import java.util.HashMap;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -30,12 +25,13 @@ public class ArgumentTokenizerTest {
         assertEquals("value1", argumentMultimap.get("key1"));
         assertEquals("value 2", argumentMultimap.get("key2"));
         assertEquals("value value value 3", argumentMultimap.get("key3"));
-
-        // TODO needs further improvement
-//        input = "233333333333/ 1 suibianshenme /key";
-//        argumentMultimap = ArgumentTokenizer.getArgumentMultimap(input);
-//        assertEquals(false, argumentMultimap.containsKey("1"));
-//        assertEquals("", argumentMultimap.get("key"));
+        /*
+        TODO needs further improvement
+        input = "233333333333/ 1 suibianshenme /key";
+        argumentMultimap = ArgumentTokenizer.getArgumentMultimap(input);
+        assertEquals(false, argumentMultimap.containsKey("1"));
+        assertEquals("", argumentMultimap.get("key"));
+        */
     }
 
 }
