+++ src/test/java/tasks/LastTest.java
@@ -1,5 +1,6 @@
 package tasks;
 
+import model.tasks.Last;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
