+++ src/test/java/tasks/ToDoTest.java
@@ -1,5 +1,6 @@
 package tasks;
 
+import model.tasks.ToDo;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
